Rename plotting.py to display.py
